+++ next-frontend/src/lib/wca/data/continents.ts
@@ -1,0 +1,14 @@
+import _ from "lodash";
+import continentDataRaw from "../../staticData/continents.json";
+
+// ----- CONTINENTS -----
+
+const fictionalContinentIds = ["_Multiple Continents"];
+const realContinents = continentDataRaw.filter(
+  (c) => !fictionalContinentIds.includes(c.id),
+);
+
+export const continents = {
+  byId: _.keyBy(continentDataRaw, "id"),
+  real: realContinents,
+};
